small cleanup

+++ include/common/type_definitions.h
@@ -61,7 +61,7 @@ using PackOfFrames = std::unordered_map<KinectId, KinectData>;
 class PackOfFramesHandler {
 public:
     PackOfFramesHandler(std::uint64_t maxDistBetweenFramesInBatch,
-                        std::uint8_t numberOfKinects,
+                        std::size_t numberOfKinects,
                         std::size_t minNumberOfFramesInPackageToAccept);
 
     void putFrame(const KinectId& kinectId, KinectData&& data);
@@ -76,7 +76,7 @@ private:
     std::map<std::uint64_t, PackOfFrames> packages;
     std::queue<PackOfFrames> readyPackOfFrames;
     std::uint64_t maxDistBetweenFramesInBatch;
-    std::uint8_t numberOfKinects;
+    std::size_t numberOfKinects;
     std::size_t minNumberOfFramesInPackageToAccept;
 };
